move out in separate file + 1 test
